Extract QuestionPanel.
