+++ lib/widgets/common/question_popover.dart
@@ -1,5 +1,5 @@
 import 'package:einbuergerungstest/services/question_database/models/question.dart';
-import 'package:einbuergerungstest/widgets/common/buttons/question_favorite_button.dart';
+import 'package:einbuergerungstest/widgets/common/question_panel.dart';
 import 'package:flutter/material.dart';
 
 class QuestionPopover extends StatelessWidget {
@@ -15,8 +15,13 @@ class QuestionPopover extends StatelessWidget {
     return Dialog(
       child: Padding(
         padding: const EdgeInsets.all(8.0),
-        child: QuestionPanel(
-          question: question,
+        child: DefaultTextStyle(
+          style: Theme.of(context).textTheme.headline6!.copyWith(
+                fontWeight: FontWeight.w400,
+              ),
+          child: QuestionPanel(
+            question: question,
+          ),
         ),
       ),
     );
@@ -34,70 +39,6 @@ class QuestionPopover extends StatelessWidget {
       context: context,
       useSafeArea: false,
       builder: (context) => QuestionPopover(question: question),
-    );
-  }
-}
-
-class QuestionPanel extends StatelessWidget {
-  const QuestionPanel({
-    required this.question,
-    this.answerIndex,
-    Key? key,
-  }) : super(key: key);
-
-  final Question question;
-  final int? answerIndex;
-
-  @override
-  Widget build(BuildContext context) {
-    return LayoutBuilder(
-      builder: (context, constraints) => Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Wrap(
-            runSpacing: 8,
-            children: [
-              if (question.hasImage)
-                Image.asset(
-                  question.imagePath!,
-                  height: constraints.maxHeight == double.infinity ? 200 : constraints.maxHeight / 6,
-                ),
-              Row(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Expanded(
-                    child: Text(
-                      question.question,
-                      style: DefaultTextStyle.of(context).style.copyWith(
-                            fontWeight: FontWeight.w600,
-                          ),
-                    ),
-                  ),
-                  SizedBox(
-                    width: 24,
-                    height: 24,
-                    child: QuestionFavoriteButton(
-                      question: question,
-                      alignment: Alignment.topRight,
-                    ),
-                  ),
-                ],
-              ),
-            ],
-          ),
-          const SizedBox(height: 8),
-          for (var i = 0; i < question.answers.length; i++) ...[
-            if (i != 0) const SizedBox(height: 2),
-            Text(
-              question.answers[i],
-              style: TextStyle(
-                color: i == question.correctAnswerIndex ? Colors.green : (i == answerIndex ? Colors.red : null),
-              ),
-            ),
-          ],
-        ],
-      ),
     );
   }
 }
